requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
protobuf==4.22.1
regex==2023.3.23
requests==2.28.2
tiktoken==0.3.2
urllib3==1.26.15
